story-teller/requirements.txt
requests
transformers
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cpu
fastapi 
uvicorn
python-dotenv
soundfile
datasets
sentencepiece